small cleanup/organization
